requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
crypto==1.4.1
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
gunicorn==20.0.4
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Naked==0.1.31
PyJWT==1.7.1
pylint==2.5.3
pyodbc==4.0.30
PyYAML==5.3.1
requests==2.24.0
shellescape==3.8.1
six==1.15.0
sql-server.pyodbc==1.0
SQLAlchemy==1.3.18
toml==0.10.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
